Moved some tests classes which were in a wrong test package.
